@@ -1,22 +1,20 @@
 <template>
   <custom-header></custom-header>
-  <router-view v-slot="slotProps">
-    <transition name="route" mode="out-in">
-      <component :is="slotProps.Component"/>
+  <router-view v-slot='slotProps'>
+    <transition name='route' mode='out-in'>
+      <component :is='slotProps.Component' />
     </transition>
   </router-view>
 </template>
 
-<script>
-import CustomHeader from "@/components/custom/CustomHeader";
+<script setup lang='ts'>
 
-export default {
-  name: 'App',
-  components: {CustomHeader},
-  created() {
-    this.$store.dispatch('auth/checkExistUser', 'logIn')
-  }
-}
+import { useStore } from 'vuex'
+import CustomHeader from '@/components/custom/CustomHeader.vue'
+
+const store = useStore()
+
+store.dispatch('auth/checkExistUser', 'logIn')
 </script>
 
 <style>
