@@ -1,6 +1,6 @@
 <template>
   <section>
-    <custom-dialog :show="isError" @close="closeModal" title="An error occurred!">
+    <custom-dialog :show='isError' @close='closeModal' title='An error occurred!'>
       <p>{{ errorMessage }}</p>
     </custom-dialog>
     <article>
@@ -8,10 +8,10 @@
         <header>
           <h2>Here you see requests</h2>
         </header>
-        <custom-spinner v-if="isLoading"/>
-        <ul v-else-if="hasRequests">
-          <request-item v-for="request in requests" :key="request.id" :email="request.email"
-                        :message="request.message"/>
+        <custom-spinner v-if='isLoading' />
+        <ul v-else-if='hasRequests'>
+          <request-item v-for='request in requests' :key='request.id' :email='request.email'
+                        :message='request.message' />
         </ul>
         <h3 v-else>You have received any requests yet!</h3>
       </custom-card>
@@ -19,41 +19,38 @@
   </section>
 </template>
 
-<script>
-import CustomCard from "@/components/custom/CustomCard";
-import RequestItem from "@/components/requests/RequestItem";
-import CustomSpinner from "@/components/custom/CustomSpinner";
-import CustomDialog from "@/components/custom/CustomDialog";
+<script setup lang='ts'>
+import CustomCard from '@/components/custom/CustomCard.vue'
+import RequestItem from '@/components/requests/RequestItem.vue'
+import CustomSpinner from '@/components/custom/CustomSpinner.vue'
+import CustomDialog from '@/components/custom/CustomDialog.vue'
+import { useStore } from 'vuex'
+import { computed } from 'vue'
 
-export default {
-  name: 'Requests',
-  components: {CustomDialog, CustomSpinner, RequestItem, CustomCard},
-  created() {
-    this.$store.dispatch('requests/setRequestsAction')
-  },
-  computed: {
-    hasRequests() {
-      return this.$store.getters['requests/hasRequests']
-    },
-    requests() {
-      return this.$store.getters['requests/requests']
-    },
-    isLoading() {
-      return this.$store.getters['requests/isLoading']
-    },
-    isError() {
-      return this.$store.getters['requests/isError']
-    },
-    errorMessage() {
-      return this.$store.getters['requests/errorMessage']
-    }
-  },
-  methods: {
-    closeModal() {
-      this.$store.commit('requests/updateError', false)
-    }
-  }
+const store = useStore()
+store.dispatch('requests/setRequestsAction')
+
+
+const hasRequests = computed(() => {
+  return store.getters['requests/hasRequests']
+})
+const requests = computed(() => {
+  return store.getters['requests/requests']
+})
+const isLoading = computed(() => {
+  return store.getters['requests/isLoading']
+})
+const isError = computed(() => {
+  return store.getters['requests/isError']
+})
+const errorMessage = computed(() => {
+  return store.getters['requests/errorMessage']
+})
+
+function closeModal() {
+  store.commit('requests/updateError', false)
 }
+
 </script>
 
 <style scoped>
